[FIX]: update  import to reflect rebased refactorings of module tree
